+++ src/Packages.php
@@ -72,10 +72,18 @@ class Packages {
 		// Proxies "activated_plugin" hook for embedded packages listen on WC plugin activation
 		// https://github.com/woocommerce/woocommerce/issues/28697.
 		if ( is_admin() ) {
-			$woocommerce_activated_plugin = get_transient( 'woocommerce_activated_plugin' );
-			if ( $woocommerce_activated_plugin ) {
+			$activated_plugin = get_transient( 'woocommerce_activated_plugin' );
+			if ( $activated_plugin ) {
 				delete_transient( 'woocommerce_activated_plugin' );
-				do_action( 'woocommerce_activated_plugin', $woocommerce_activated_plugin );
+
+				/**
+				 * WooCommerce is activated hook.
+				 *
+				 * @since 5.0.0
+				 * @param bool $activated_plugin Activated plugin path,
+				 *                               generally woocommerce/woocommerce.php.
+				 */
+				do_action( 'woocommerce_activated_plugin', $activated_plugin );
 			}
 		}
 	}